# rest-server/doc/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "DM124 REST API",
    "description": "Trabalho Prático da disciplina DM124 - Lucas Leon Labanca"
  },
  "host": "localhost:8080",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic",
      "in": "header",
      "name": "Authorization",
      "description": "Basic Auth Base 64"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/questions": {
      "post": {
        "tags": [
          "Questions"
        ],
        "description": "Add a question",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Question",
            "in": "body",
            "description": "Question",
            "required": true,
            "schema": {
              "$ref": "#/definitions/QuestionToAdd"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Question added succesfully",
            "schema": {
              "$ref": "#/definitions/QuestionAdded"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Questions"
        ],
        "description": "Get a list of all of the questions",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/QuestionList"
            },
            "description": "Question list found"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/questions/{questionId}": {
      "get": {
        "tags": [
          "Questions"
        ],
        "description": "Get a question by ID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Question ID"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Question"
            },
            "description": "Question found"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Questions"
        ],
        "description": "Update a question",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Question ID"
          },
          {
            "name": "Question",
            "in": "body",
            "description": "Question",
            "required": true,
            "schema": {
              "$ref": "#/definitions/QuestionToAdd"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/QuestionAdded"
            },
            "description": "Question updated succesfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Questions"
        ],
        "description": "Delete a question by ID",
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Question ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Question deleted"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/answers": {
      "post": {
        "tags": [
          "Answers"
        ],
        "description": "Add an answer",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Answer",
            "in": "body",
            "description": "Answer",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AnswerToAdd"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Answer added succesfully",
            "schema": {
              "$ref": "#/definitions/AnswerAdded"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Answers"
        ],
        "description": "Get a list of all of the answers",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/AnswerList"
            },
            "description": "Answer list found"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/answers/{answerId}": {
      "get": {
        "tags": [
          "Answers"
        ],
        "description": "Get an answer by ID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "answerId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Answer ID"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Answer"
            },
            "description": "Answer found"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Answers"
        ],
        "description": "Update an answer",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "answerId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Answer ID"
          },
          {
            "name": "Answer",
            "in": "body",
            "description": "Answer",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AnswerToAdd"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/AnswerAdded"
            },
            "description": "Answer updated succesfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Answers"
        ],
        "description": "Delete an answer by ID",
        "parameters": [
          {
            "name": "answerId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Answer ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Answer deleted"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
    "Question": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "status": {
          "type": "string",
          "example": "New"
        },
        "description": {
          "type": "string",
          "example": "Question 1"
        },
        "createdAt": {
          "type": "string",
          "example": "2021-09-24T00:52:02.775Z"
        },
        "updatedAt": {
          "type": "string",
          "example": "2021-09-24T00:52:02.775Z"
        },
        "answer": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number",
              "example": 1
            },
            "key": {
              "type": "number",
              "example": 1
            },
            "name": {
              "type": "string",
              "example": "Days of week answer"
            },
            "answer": {
              "type": "string",
              "example": "Answer to Question 1"
            },
            "createdAt": {
              "type": "string",
              "example": "2021-09-24T00:53:58.517Z"
            },
            "updatedAt": {
              "type": "string",
              "example": "2021-09-24T00:53:58.517Z"
            },
            "questionId": {
              "type": "number",
              "example": 1
            }
          }
        },
        "options": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 3
              },
              "option": {
                "type": "string",
                "example": "Option-3"
              },
              "questionId": {
                "type": "number",
                "example": 1
              }
            }
          }
        }
      }
    },
    "QuestionList": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "status": {
            "type": "string",
            "example": "New"
          },
          "description": {
            "type": "string",
            "example": "Question 1"
          },
          "createdAt": {
            "type": "string",
            "example": "2021-09-24T00:52:02.775Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2021-09-24T00:52:02.775Z"
          },
          "answer": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 1
              },
              "key": {
                "type": "number",
                "example": 1
              },
              "name": {
                "type": "string",
                "example": "Days of week answer"
              },
              "answer": {
                "type": "string",
                "example": "Answer to Question 1"
              },
              "createdAt": {
                "type": "string",
                "example": "2021-09-24T00:53:58.517Z"
              },
              "updatedAt": {
                "type": "string",
                "example": "2021-09-24T00:53:58.517Z"
              },
              "questionId": {
                "type": "number",
                "example": 1
              }
            }
          },
          "options": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 3
                },
                "option": {
                  "type": "string",
                  "example": "Option-3"
                },
                "questionId": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        }
      }
    },
    "QuestionToAdd": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "example": "Question 1"
        },
        "status": {
          "type": "string",
          "example": "New"
        },
        "options": {
          "type": "array",
          "example": [
            "Option-1",
            "Option-2",
            "Option-3"
          ],
          "items": {
            "type": "string"
          }
        }
      }
    },
    "QuestionAdded": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "status": {
          "type": "string",
          "example": "New"
        },
        "description": {
          "type": "string",
          "example": "Question 1"
        },
        "updatedAt": {
          "type": "string",
          "example": "2021-09-24T00:52:02.775Z"
        },
        "createdAt": {
          "type": "string",
          "example": "2021-09-24T00:52:02.775Z"
        }
      }
    },
    "Answer": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "key": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Days of week answer"
        },
        "answer": {
          "type": "string",
          "example": "Answer to Question 2"
        },
        "createdAt": {
          "type": "string",
          "example": "2021-09-22T03:42:40.384Z"
        },
        "updatedAt": {
          "type": "string",
          "example": "2021-09-22T03:43:36.138Z"
        },
        "questionId": {
          "type": "number",
          "example": 1
        },
        "question": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number",
              "example": 1
            },
            "status": {
              "type": "string",
              "example": "New"
            },
            "description": {
              "type": "string",
              "example": "Question 1"
            },
            "createdAt": {
              "type": "string",
              "example": "2021-09-22T03:39:44.484Z"
            },
            "updatedAt": {
              "type": "string",
              "example": "2021-09-22T03:39:44.484Z"
            }
          }
        }
      }
    },
    "AnswerList": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "key": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Days of week answer"
          },
          "answer": {
            "type": "string",
            "example": "Answer to Question 2"
          },
          "createdAt": {
            "type": "string",
            "example": "2021-09-22T03:42:40.384Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2021-09-22T03:43:36.138Z"
          },
          "questionId": {
            "type": "number",
            "example": 1
          },
          "question": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 1
              },
              "status": {
                "type": "string",
                "example": "New"
              },
              "description": {
                "type": "string",
                "example": "Question 1"
              },
              "createdAt": {
                "type": "string",
                "example": "2021-09-22T03:39:44.484Z"
              },
              "updatedAt": {
                "type": "string",
                "example": "2021-09-22T03:39:44.484Z"
              }
            }
          }
        }
      }
    },
    "AnswerToAdd": {
      "type": "object",
      "properties": {
        "key": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Days of week answer"
        },
        "answer": {
          "type": "string",
          "example": "Answer to Question 1"
        },
        "questionId": {
          "type": "number",
          "example": 1
        }
      }
    },
    "AnswerAdded": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "key": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Days of week answer"
        },
        "answer": {
          "type": "string",
          "example": "Answer to Question 1"
        },
        "questionId": {
          "type": "number",
          "example": 1
        },
        "updatedAt": {
          "type": "string",
          "example": "2021-09-23T03:33:23.630Z"
        },
        "createdAt": {
          "type": "string",
          "example": "2021-09-23T03:33:23.630Z"
        }
      }
    },
...
  }
}